@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 class DetailRow extends StatefulWidget {
   final String title;
   final String value;
-  DetailRow({@required this.title, this.value});
+  DetailRow({required this.title, required this.value});
 
   @override
   _DetailRowState createState() => _DetailRowState();
@@ -14,7 +14,6 @@ class _DetailRowState extends State<DetailRow> {
   Widget build(BuildContext context) {
     final width = MediaQuery.of(context).size.width;
     return ListTile(
-      // mainAxisAlignment: MainAxisAlignment.spaceBetween,
       title: Text(widget.title),
       trailing: SizedBox(
           width: width / 2,
@@ -22,19 +21,7 @@ class _DetailRowState extends State<DetailRow> {
             widget.value,
             textAlign: TextAlign.right,
           )),
-      // contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: -4),
       visualDensity: VisualDensity.compact.copyWith(vertical: -4),
-
-      // children: [
-      //   SizedBox(
-      //     width: 200,
-      //     child: Text(title),
-      //   ),
-      //   SizedBox(
-      //     width: width - 238,
-      //     child: Text(value ?? 'null', textAlign: TextAlign.end),
-      //   ),
-      // ],
     );
   }
 }
