@@ -7,7 +7,6 @@ import 'package:adapty_flutter/models/adapty_error.dart';
 import 'package:adapty_flutter/models/adapty_paywall.dart';
 import 'package:adapty_flutter/models/adapty_profile.dart';
 import 'package:adapty_flutter/models/adapty_purchaser_info.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
 import 'constants/arguments_names.dart';
@@ -49,57 +48,66 @@ class Adapty {
     return AdaptyLogLevel.values[index];
   }
 
-  static Future<bool> identify(String customerUserId) {
-    return _invokeMethodHandlingErrors<bool>(Method.identify, {
+  static Future<bool> identify(String customerUserId) async {
+    final result = await _invokeMethodHandlingErrors<bool>(Method.identify, {
       Argument.customerUserId: customerUserId,
     });
+    return result ?? false;
   }
 
   static Future<GetPaywallsResult> getPaywalls({bool forceUpdate = false}) async {
-    final result = await _invokeMethodHandlingErrors<String>(Method.getPaywalls, {
+    final result = (await _invokeMethodHandlingErrors<String>(Method.getPaywalls, {
       Argument.forceUpdate: forceUpdate,
-    });
+    })) as String;
     return GetPaywallsResult.fromMap(json.decode(result));
   }
 
   static Future<AdaptyPurchaserInfo> getPurchaserInfo({bool forceUpdate = false}) async {
-    final result = await _invokeMethodHandlingErrors<String>(Method.getPurchaserInfo, {
+    final result = (await _invokeMethodHandlingErrors<String>(Method.getPurchaserInfo, {
       Argument.forceUpdate: forceUpdate,
-    });
+    })) as String;
     return AdaptyPurchaserInfo.fromMap(json.decode(result));
   }
 
-  static Future<bool> updateProfile(AdaptyProfileParameterBuilder builder) {
-    return _invokeMethodHandlingErrors<bool>(Method.updateProfile, {
+  static Future<bool> updateProfile(AdaptyProfileParameterBuilder builder) async {
+    final result = await _invokeMethodHandlingErrors<bool>(Method.updateProfile, {
       Argument.params: builder.map,
     });
+    return result ?? false;
   }
 
   static Future<MakePurchaseResult> makePurchase(AdaptyProduct product) async {
-    final result = await _invokeMethodHandlingErrors<String>(Method.makePurchase, {
+    final result = (await _invokeMethodHandlingErrors<String>(Method.makePurchase, {
       Argument.productId: product.vendorProductId,
       if (product.variationId != null) Argument.variationId: product.variationId,
-    });
+    })) as String;
     return MakePurchaseResult.fromJson(json.decode(result));
   }
 
   static Future<RestorePurchasesResult> restorePurchases() async {
-    final result = await _invokeMethodHandlingErrors<String>(Method.restorePurchases);
+    final result = (await _invokeMethodHandlingErrors<String>(Method.restorePurchases)) as String;
     return RestorePurchasesResult.fromJson(json.decode(result));
   }
 
-  static Future<bool> updateAttribution(Map attribution, {@required AdaptyAttributionNetwork source, String networkUserId}) {
-    return _invokeMethodHandlingErrors<bool>(Method.updateAttribution, {
+  static Future<bool> updateAttribution(Map attribution, {required AdaptyAttributionNetwork source, String? networkUserId}) async {
+    final result = await _invokeMethodHandlingErrors<bool>(Method.updateAttribution, {
       Argument.attribution: attribution,
       Argument.source: source.stringValue(),
       if (networkUserId != null) Argument.networkUserId: networkUserId,
     });
+    return result ?? false;
   }
 
-  static Future<void> logShowPaywall({@required AdaptyPaywall paywall}) {
-    return _invokeMethodHandlingErrors<void>(Method.logShowPaywall, {
-      Argument.variationId: paywall.variationId,
-    });
+  static Future<void> logShowPaywall({required AdaptyPaywall paywall}) async {
+    if (Platform.isIOS) {
+      await _invokeMethodHandlingErrors<void>(Method.logShowPaywall, {
+        Argument.variationId: paywall.variationId,
+      });
+    } else {
+      _invokeMethodHandlingErrors<void>(Method.logShowPaywall, {
+        Argument.variationId: paywall.variationId,
+      });
+    }
   }
 
   static Future<void> setExternalAnalyticsEnabled(bool enabled) {
@@ -115,28 +123,29 @@ class Adapty {
     });
   }
 
-  static Future<bool> logout() {
-    return _invokeMethodHandlingErrors<bool>(Method.logout);
+  static Future<bool> logout() async {
+    final result = await _invokeMethodHandlingErrors<bool>(Method.logout);
+    return result ?? false;
   }
 
   // ––––––– IOS ONLY METHODS –––––––
 
   static Future<void> setApnsToken(String token) {
-    if (!Platform.isIOS) return null;
+    if (!Platform.isIOS) return Future.value();
     return _invokeMethodHandlingErrors<void>(Method.setApnsToken, {Argument.value: token});
   }
 
   static Future<void> handlePushNotification(Map userInfo) {
-    if (!Platform.isIOS) return null;
+    if (!Platform.isIOS) return Future.value();
     return _invokeMethodHandlingErrors<void>(Method.handlePushNotification, {Argument.userInfo: userInfo});
   }
 
   static Future<void> setFallbackPaywalls(String paywalls) {
-    if (!Platform.isIOS) return null;
+    if (!Platform.isIOS) return Future.value();
     return _invokeMethodHandlingErrors<void>(Method.setFallbackPaywalls, {Argument.paywalls: paywalls});
   }
 
-  static Future<AdaptyResult> validateReceipt(String receipt) async {
+  static Future<AdaptyResult?> validateReceipt(String receipt) async {
     if (!Platform.isIOS) return null;
 
     try {
@@ -150,49 +159,49 @@ class Adapty {
     }
   }
 
-  static Future<AdaptyPromo> getPromo() async {
+  static Future<AdaptyPromo?> getPromo() async {
     if (!Platform.isIOS) return null;
 
-    final String result = await _invokeMethodHandlingErrors(Method.getPromo);
+    final String? result = await _invokeMethodHandlingErrors(Method.getPromo);
     if (result != null) {
       return AdaptyPromo.fromJson(jsonDecode(result));
     }
     return null;
   }
 
-  static Future<MakePurchaseResult> makeDeferredPurchase(String productId) async {
+  static Future<MakePurchaseResult?> makeDeferredPurchase(String productId) async {
     if (!Platform.isIOS) return null;
 
-    final String result = await _invokeMethodHandlingErrors(Method.makeDeferredPurchase, {
+    final String result = await (_invokeMethodHandlingErrors(Method.makeDeferredPurchase, {
       Argument.productId: productId,
-    });
+    }) as FutureOr<String>);
     return MakePurchaseResult.fromJson(json.decode(result));
   }
 
   static Future<void> presentCodeRedemptionSheet() {
-    if (!Platform.isIOS) return null;
+    if (!Platform.isIOS) return Future.value();
     return _invokeMethodHandlingErrors<void>(Method.presentCodeRedemptionSheet);
   }
 
   // ––––––– ANDROID ONLY METHODS –––––––
 
   static Future<bool> newPushToken(String token) async {
-    if (!Platform.isAndroid) return null;
+    if (!Platform.isAndroid) return false;
 
-    final bool result = await _invokeMethodHandlingErrors(Method.newPushToken, {Argument.pushToken: token});
-    return result;
+    final bool? result = await _invokeMethodHandlingErrors(Method.newPushToken, {Argument.pushToken: token});
+    return result ?? false;
   }
 
   static Future<bool> pushReceived(Map<String, String> message) async {
-    if (!Platform.isAndroid) return null;
+    if (!Platform.isAndroid) return false;
 
-    final bool result = await _invokeMethodHandlingErrors(Method.pushReceived, {Argument.pushMessage: message});
-    return result;
+    final bool? result = await _invokeMethodHandlingErrors(Method.pushReceived, {Argument.pushMessage: message});
+    return result ?? false;
   }
 
   // ––––––– INTERNAL –––––––
 
-  static Future<T> _invokeMethodHandlingErrors<T>(String method, [dynamic arguments]) async {
+  static Future<T?> _invokeMethodHandlingErrors<T>(String method, [dynamic arguments]) async {
     try {
       return await _channel.invokeMethod<T>(method, arguments);
     } on PlatformException catch (e) {
@@ -203,23 +212,25 @@ class Adapty {
   static Future<dynamic> _handleIncomingMethodCall(MethodCall call) {
     switch (call.method) {
       case Method.deferredPurchaseProduct:
-        var productIdentifier = call.arguments as String;
-        _deferredPurchasesController.add(productIdentifier);
-        return null;
+        var productIdentifier = call.arguments as String?;
+        if (productIdentifier != null) {
+          _deferredPurchasesController.add(productIdentifier);
+        }
+        return Future.value(null);
       case Method.getPaywallsResult:
         var result = call.arguments as String;
         _getPaywallsResultController.add(GetPaywallsResult.fromMap(json.decode(result)));
-        return null;
+        return Future.value(null);
       case Method.purchaserInfoUpdate:
         var result = call.arguments as String;
         _purchaserInfoUpdateController.add(AdaptyPurchaserInfo.fromMap(json.decode(result)));
-        return null;
+        return Future.value(null);
       case Method.promoReceived:
         var result = call.arguments as String;
         _promosReceiveController.add(AdaptyPromo.fromJson(json.decode(result)));
-        return null;
+        return Future.value(null);
       default:
-        return null;
+        return Future.value(null);
     }
   }
 }
