@@ -3,16 +3,16 @@ import '../helpers/extensions.dart';
 
 class AdaptyPromo {
   /// The type of the promo offer.
-  final String promoType;
+  final String? promoType;
 
   /// The identifier of the variation, used to attribute purchases to the promo.
-  final String variationId;
+  final String? variationId;
 
   /// The time when the current promo offer will expire.
-  final DateTime expiresAt;
+  final DateTime? expiresAt;
 
   /// A [AdaptyPaywall] object.
-  final AdaptyPaywall paywall;
+  final AdaptyPaywall? paywall;
 
   AdaptyPromo.fromJson(Map<String, dynamic> json)
       : promoType = json[_Keys.promoType],
