@@ -46,21 +46,21 @@ class _MainScreenState extends State<MainScreen> {
   }
 
   void _subscribeForStreams(BuildContext context) {
-    final scaffoldState = scaffoldKey.currentState as ScaffoldState;
+    final scaffoldMessenger = ScaffoldMessenger.of(context);
     Adapty.purchaserInfoUpdateStream.listen((purchaserInfo) {
-      scaffoldState?.showSnackBar(buildPurchaserInfoSnackbar(context, purchaserInfo));
+      scaffoldMessenger.showSnackBar(buildPurchaserInfoSnackbar(context, purchaserInfo));
       print('#Example# purchaserInfoUpdateStream:\n $purchaserInfo');
     });
     Adapty.promosReceiveStream.listen((promo) {
-      scaffoldState?.showSnackBar(buildPromoSnackbar(context, promo));
+      scaffoldMessenger.showSnackBar(buildPromoSnackbar(context, promo));
       print('#Example# promosReceiveStream:\n $promo');
     });
     Adapty.deferredPurchasesStream.listen((event) {
-      scaffoldState?.showSnackBar(buildSimpleSnackbar(event));
+      scaffoldMessenger.showSnackBar(buildSimpleSnackbar(event));
       print('#Example# deferredPurchasesStream:\n $event');
     });
     Adapty.getPaywallsResultStream.listen((event) {
-      scaffoldState?.showSnackBar(buildSimpleSnackbar('Paywalls Updated!'));
+      scaffoldMessenger.showSnackBar(buildSimpleSnackbar('Paywalls Updated!'));
       print('#Example# getPaywallsResultStream:\n $event');
     });
   }
@@ -90,11 +90,11 @@ class _MainScreenState extends State<MainScreen> {
         'Identify',
         () async {
           callAdaptyMethod(() async {
-            final scaffoldState = scaffoldKey.currentState as ScaffoldState;
+            final scaffoldMessenger = ScaffoldMessenger.of(context);
             final installId = await Service.getOrCreateInstallId();
             bool res = false;
             res = await Adapty.identify(installId);
-            scaffoldState.showSnackBar(buildSimpleSnackbar(res ? 'You identify with $installId.' : 'You were not identified.'));
+            scaffoldMessenger.showSnackBar(buildSimpleSnackbar(res ? 'You identify with $installId.' : 'You were not identified.'));
           });
         },
         openNewScreen: false,
@@ -135,10 +135,9 @@ class _MainScreenState extends State<MainScreen> {
       _buildMethodTile('Restore Purchases', () {
         callAdaptyMethod(() async {
           final res = await Adapty.restorePurchases();
-          if (res != null) {
-            final state = scaffoldKey.currentState as ScaffoldState;
-            state.showSnackBar(buildPurchaserInfoSnackbar(context, res.purchaserInfo));
-          }
+
+          final state = ScaffoldMessenger.of(context);
+          state.showSnackBar(buildPurchaserInfoSnackbar(context, res.purchaserInfo));
         });
       }),
       _buildMethodTile(
@@ -158,8 +157,7 @@ class _MainScreenState extends State<MainScreen> {
             'af_status': 'Organic',
             'is_first_launch': false,
           };
-
-          await Adapty.updateAttribution(testAttributionMap, source: AdaptyAttributionNetwork.appsflyer);
+          await Adapty.updateAttribution(testAttributionMap, source: AdaptyAttributionNetwork.adjust);
           print('#Example# updateAttribution done!');
         }),
         openNewScreen: false,
@@ -209,8 +207,12 @@ class _MainScreenState extends State<MainScreen> {
       _buildMethodTile(
         'Logout',
         () => callAdaptyMethod(() async {
-          await Adapty.logout();
-          print('#Example# logout done!');
+          final result = await Adapty.logout();
+          if (result) {
+            print('#Example# logout done!');
+          } else {
+            print('#Example# logout is not succes');
+          }
         }),
       ),
     ];
@@ -234,10 +236,10 @@ class _MainScreenState extends State<MainScreen> {
   Widget _buildInstallIdTile() {
     return ListTile(
       title: Text('Your Install Id (generated by app)'),
-      subtitle: FutureBuilder(
+      subtitle: FutureBuilder<String>(
         future: Service.getOrCreateInstallId(),
         builder: (ctx, snapshot) {
-          return Text(snapshot.hasData ? snapshot.data : '');
+          return Text(snapshot.data ?? '');
         },
       ),
     );
@@ -246,11 +248,12 @@ class _MainScreenState extends State<MainScreen> {
   Widget _buildLogLevelTile() {
     return ListTile(
       title: Text('Log Level'),
-      subtitle: FutureBuilder(
+      subtitle: FutureBuilder<AdaptyLogLevel>(
         future: Adapty.getLogLevel(),
         builder: (ctx, snapshot) {
           if (snapshot.hasError && snapshot.error is AdaptyError) {
-            AdaptyErrorDialog.showAdaptyErrorDialog(context, snapshot.error);
+            final adaptyError = snapshot.error as AdaptyError;
+            AdaptyErrorDialog.showAdaptyErrorDialog(context, adaptyError);
           }
           return Text(snapshot.hasData ? snapshot.data.toString() : '');
         },
