@@ -1,4 +1,5 @@
 import 'package:adapty_flutter/models/adapty_purchaser_info.dart';
+import 'package:adapty_flutter_example/Helpers/value_to_string.dart';
 import 'package:adapty_flutter_example/screens/access_levels_screen.dart';
 import 'package:adapty_flutter_example/screens/non_subscriptions_screen.dart';
 import 'package:adapty_flutter_example/screens/subscriptions_screen.dart';
@@ -15,40 +16,34 @@ class PurchaserInfoScreen extends StatefulWidget {
 class _PurchaserInfoScreenState extends State<PurchaserInfoScreen> {
   @override
   Widget build(BuildContext context) {
-    // final details = {
-    //   'Data State': result.dataState == AdaptyDataState.cached ? 'cached' : 'synced',
-    //   'Customer User Id': valueToString(purchaserInfo.customerUserId),
-    // };
-    final detailPages = widget.purchaserInfo != null
-        ? {
-            'Access Levels': () => Navigator.of(context).push(MaterialPageRoute(builder: (ctx) => AccessLevelsScreen(widget.purchaserInfo.accessLevels))),
-            'Subscriptions': () => Navigator.of(context).push(MaterialPageRoute(builder: (ctx) => SubscriptionsScreen(widget.purchaserInfo.subscriptions))),
-            'Non Subscriptions': () => Navigator.of(context).push(MaterialPageRoute(builder: (ctx) => NonSubscriptionsScreen(widget.purchaserInfo.nonSubscriptions))),
-          }
-        : null;
+    final details = {
+      'Customer User Id': valueToString(widget.purchaserInfo.customerUserId),
+    };
+
+    final detailPages = {
+      'Access Levels': () => Navigator.of(context).push(MaterialPageRoute(builder: (ctx) => AccessLevelsScreen(widget.purchaserInfo.accessLevels))),
+      'Subscriptions': () => Navigator.of(context).push(MaterialPageRoute(builder: (ctx) => SubscriptionsScreen(widget.purchaserInfo.subscriptions))),
+      'Non Subscriptions': () => Navigator.of(context).push(MaterialPageRoute(builder: (ctx) => NonSubscriptionsScreen(widget.purchaserInfo.nonSubscriptions))),
+    };
+
     return Scaffold(
-      appBar: AppBar(
-        title: const Text('Purchaser Info'),
-        leading: IconButton(
-          icon: Icon(
-            Icons.arrow_back_ios_outlined,
-            size: 24,
-          ),
-          onPressed: Navigator.of(context).pop,
-        ),
-      ),
-      body: widget.purchaserInfo != null
-          ? ListView(
-              children: [
-                DetailsContainer(
-                  // details: details,
-                  detailPages: detailPages,
-                ),
-              ],
-            )
-          : Center(
-              child: Text('Purchaser Info was not received.'),
+        appBar: AppBar(
+          title: const Text('Purchaser Info'),
+          leading: IconButton(
+            icon: Icon(
+              Icons.arrow_back_ios_outlined,
+              size: 24,
             ),
-    );
+            onPressed: Navigator.of(context).pop,
+          ),
+        ),
+        body: ListView(
+          children: [
+            DetailsContainer(
+              details: details,
+              detailPages: detailPages,
+            ),
+          ],
+        ));
   }
 }
